Add JSON linting support
This should resolve scenarios where we are seeing linter differences
between local and the CI pipeline (because the CI pipeline is correctly
linting JSON files whereas the npm setup is not).

Eventually we need to converge on having a shared .eslintrc.json file
for all of our functions (as much as possible) and possibly using a tool
like lerna to manage the various projects that we have in the repo;
however, for now here are the steps to follow to get _other_ projects in
sync with this PoC app config:

1. Install the `eslint-plugin-json` package as a dev dependency
```
npm install --save-dev eslint-plugin-json
```

2. Add `"plugin:json/recommended"` to the `extends` section of
`.eslintrc.json`

3. Update `package.json` to have `,.json` in the `lint` and `lint:fix`
commands.