Remove the 'hello world' function from the stack.
This lambda function was part of the AWS proof of concept and has served its purpose.
This removes the function from the AWS CDK stack and removes the function implementation.
This function was occupying the GET method space in the '/portfolioDrafts' route that is needed for the implementation of function 'getPortfolioDrafts'.

Ticket: AT-6362