Improve portfolioStep validation
Technically this doesn't totally validate; it's just a type guard. The
docs have been updated to clarify that. Additionally, the code is
improved to not fail when object is null. From there, it's a trivial
addition to introduce the check the input is an object and remove a use
of `any`.