Refactor to use the Include macro
The things done in this commit won't make a whole lot of sense if you're
not familiar with CloudFormation. CloudFormation has two concepts:
functions and transforms. The two are somewhat related but not entirely
the same. So functions get interpreted directly by CloudFormation in the
template. They do things like reference parameter and resource values
(Fn::Ref) or perform string substitution (Fn::Sub) or even import values
from another stack (Fn::ImportValue). They can really only modify their
own small block of the template (basically, their parameters). There is
a special function to invoke a Transform (Fn::Transform). Transforms can
make structural modifications to the whole template. There are two
Transforms that are provided directly by AWS but you can also make your
own. We care about AWS::Include here. AWS::Include lets you take a file
in S3 and basically copy-and-paste it directly into your template. That
file needs to be valid CloudFormation. If it is, the resources and
values in that snippet get added into that location in your template. If
that snippet invokes more functions, those get resolved.

This means that we can add direct calls to Fn::Sub or other resources in
our OpenAPI spec. So long as we use the long form of the functions
(Fn::NAME) instead of the short form (!NAME) our file is also still a
valid OpenAPI Specification that can be read by Swagger Hub or whatever
other tool we need.

Now a final important thing is that to reference the values or
attributes of a resource within a Fn::Sub (or even Ref or any other),
the need to have predictable logical IDs. By default, the CDK generates
fairly long and generally unpredictable names for resources. It does
give you an escape hatch though using `overrideLogicalId()` on the
underlying L1 construct. Since these only need to be unique within a
stack and the underlying physical ID (here, the name of the Lambda
Function itself) is still mostly globally unique, we should be totally
safe.

This means that we only need to create our Lambda functions, give them
the Logical IDs that are expected by the OpenAPI Specification document,
and then use the SpecRestApi resource and we'll be good to go for using
the built-in API Gateway validation for resources, guaranteeing (mostly)
that by the time they get to our functions, they are structurally valid.

There is still quite a bit to do to clean this up. For the sake of
demonstration and simplicity, there is only one route that has been
added here so all the others are going to need to be added back. To
reduce the amount of Copy/Paste that goes on, we probably still want to
use the ApiFunction construct and its subclasses which is going to
require some modifications there, and we're going to need to make sure
this all still works for the /taskorder/ routes or come up with another
way to handle them if it doesn't.