Make an SSM parameter with User Pool ID available
To support the UI repo's CDK stack to be able to create a User Pool
Client, we need to actually make the pool's ID available somewhere. This
does a few things:

1. Renames `TicketId` to `EnvironmentId` to support names like 'dev',
'staging', or 'prod' more clearly.
2. Adds the SSM parameter for the User Pool ID with a
good-enough-for-now naming convention until we have time to formalize a
process for this (if that ever becomes necessary)
3. Updates existing code to use `EnvironmentId` naming