Avoid invalid uses of INVALID_INPUT
Using the INVALID_INPUT error code is supposed to be an indication that
the response is an actual ValidationError response. Unfortunately, we
have seen it used incorrectly for any time a 4xx error is being returned
and this has been inconsistent.

This introduces a change to the `utils/response.ts` to ensure that the
only time that INVALID_INPUT responses can be issued is by creating a
ValidationErrorResponse object and providing the error map and a
message. For all other error responses, OTHER will be used as is correct
per the API spec.

In a few places, this also hides internal error details from the actual
error response. Those never should have been exposed in the first place.