Normalize environment names and IDs
There are some places (like stack names) where the title-case variant of
the environment name looks and is best; however, there are some
scenarios where we need a lowercase identifier as well.

This gives us a consistent way to handle environment names and moves all
of that logic to the app level so that it is not happening in individual
stacks and constructs. Previously, we would have had problems if an
environment ID with special characters was entered but that is no longer
an issue.

Currently, we are only passing the environment ID anywhere and it is
only to one stack. The Environment Name is purely used for
CloudFormation stack names. Both are still declared and initialized so
that we can pass them in Prop structs in the future as needed. Similar
code will need to be added to the UI repo since there isn't a
particularly great way to share code between the two repositories.

Tests are added for this code; however, we have not and are not
collecting coverage data for the CDK code yet.