Update API spec YAML and interface model
Update "name" to "portfolio_name" to make the Portfolio
Draft Summary schema clear.

Include additional attributes as optional for the PortfolioDraftSumary
interface to account for the before and after creation
in the Portflio Summary. The attributes are added based on the
step (portoflio, funding, application) that is updated.

Ticket: AT-6575