Add additional tests and update model
To account for the additional attributes being added to the
portfolio draft summary, tests were added to the create funding
and portoflio step, in addition to the get portfolio draft.

Model for PortfolioDraftSummary updated to reflect the addition
of the steps and attributes as they are added to the portoflio
in the database.

Ticket: AT-6575