Add IAM permissions for each role
This creates 3 new roles intended to be used by humans directly and
1 role to be passed to CloudFormation. The three new roles are for
QA/Testing, Security Auditors, and Developers. The CloudFormation role
is for developers to pass to the `--role-arn` or equivalent flag when
building using the CDK.

Access is currently granted to the three job function roles to the
entire Central/Management account in AWS Organizations. Long-term, that
is not a good idea and we should find a way to lock it down a bit but
right now this gives us two significant benefits:

  1. Any user can assume any job role and see what those users see,
  which will be especially helpful for developers to help other roles
  troubleshoot other roles' IAM
  2. It simplifies the configuration and steps to assume the job-based
  roles compared to needing to double-hop

Most of the logic around assuming roles is going to need fairly
significant refactoring to support IAM Identity Federation in the near
future anyway (so that we are not relying on IAM Users directly). So
this feels like a pretty good interim solution but we can try to put in
the effort to lock it down if required.

The roles are created as follows:

 - QA has access to view all "safe" resources, the items in DynamoDB,
   and to GET all API Gateway restapi resources
 - Auditors have all QA access plus access to all CloudWatch Logs, and
   all actions in the SecurityAudit AWS-managed job-function policy
 - Developers have all the access of QA plus read access to all
   CloudWatch Logs, as well as permission to use the CloudFormation
   execution role, and full access to CloudFormation
 - The CloudFormation execution role has AdministratorAccess

All roles have an explicit Deny for all actions involving AWS
Organizations; however, this is not enforced by an SCP or Permissions
Boundary so we do allow new roles to be created with those permissions
(but it would have to be done with CloudFormation).

The Roles are all made available as CloudFormation stack exports. The
really cool thing about this is that it allows other stacks to grant the
roles access to specific resources that get created later (though there
are not any examples of that added in this PR since there is not really
a need).

We also make use of a pretty basic CloudFormation Custom Resource to get
the Account ID of the AWS Organizations Management Account. We are only
using one field from the output of that API call.

Finally, creation of the IAM stack at all is guarded by an environment
variable. Since we are created named IAM resources, creating duplicate
stacks would result in failures and folks have gotten accustomed to
using `--all` on the CDK CLI. This adds a warning if the stack gets
created at all to remind users that everything will get wonky if we try
to spin up two if these in the same account (actually, it'll just
totally fail to create but making it sound spookier hopefully encourages
users to pay extra attention to whether it's intentional). The only
environment where multiple stacks being created is probable is the
Sandbox environment but it's also the most commonly-used environment.

So in summary there's a lot of room for improvement here:
 - using federated identity instead of IAM users in the central account
 - limiting the permissions on the CloudFormation role
 - moving to having resources grant access to specific ATAT roles
 - find some more explicit denies
 - improve the trust process

This gets us a good start, likely gives users the majority of access
that they need for day-to-day work, and is a solid base to continue
iterating on.