Fix missing and update dependencies
We have a good chunk of dependencies that need updates so those have all
been updated. Most are dev dependencies but others include: AWS SDK, AWS
CDK, cdk-nag, and linters. We also were not installing `typescript`
specifically as a dev dependency in each project which resulted in
warnings during install.