Invoke SonarCloud in GitHub Actions
This creates a separate workflow for invoking SonarCloud. The
`pull_request` event context may not always have access to the
repository's secrets (such as when PRs come from Dependabot or external
forks). We still need SonarCloud to be invoked but we need to do so
safely and without potentially giving a malicious PR any access to our
secrets.

This is done by creating a `workflow_run`-triggered Workflow. This
workflow gets triggered when the `Build and Test API` workflow has
completed (regardless of whether it was successful or not). We then
checkout the code from the pull request, download the coverage data and
put it in place, and then invoke the SonarScanner which performs a
static analysis of the code without executing it.