Trim PortfolioDraftSummary responses correctly
This has an unfortunate requirement to create an object that maps all
the fields of an interface to null; however, it gives us a few really
cool benefits. First is that it is an exhaustive list of all the fields
of an interface (including its parent interfaces) which means that we
can trust it when using the newly-created `pick` function. It also
enforces that the list has to be updated whenever the interface gains or
removes members.

This is all required because interfaces do not exist in JavaScript and
so the information is not available at runtime. Having a data structure
whose fields are enforced at compile time with basically that type data
exposes it at run time.

This method is currently only implemented for the GetPortfolioDrafts
route; however, it is generic and can be added and adapted to other
routes over time (which we should do because it also allows us to start
storing metadata in DynamoDB that we don't want to expose via the API).

This also adds a ProjectionExpression to the DynamoDB Scan which will
limit the attributes retrieved from the database. This utilizes our new
attribute mapping. Perhaps the subsequent pick is not really needed but
it ensures that if there is extra data in the DynamoDB response that we
trim it out at the app layer.