Add Gateway Responses for more error types
Overriding DEFAULT_4XX and DEFAULT_5XX ensure that all errors returned
by the API are similar in format to those returned by the actual API
functions:

    {
      "code": "OTHER",
      "message": string
    }

A few additional responses are overriden to contextually provide (or
mask) additional information in the response. For the vast majority of
users, this means that the `/` route (and other non-existent routes) now
look the same as a 403/404 error would.