Update example of middyHandler to use portfolioStep fn
middyHandler is now an implementation of createPortfolioStep, with
a CustomApiGatewayProxyHandler wrapper included for the example (will
be moved somewhere else for the actual implementation).
This commit also includes my custom middleware that uses AJV, but its
too slow for prod (3 second start up time), so ive left a comment
and decided to use the base middy validator.