Initial AWS X-Ray integration
This performs an initial integration with AWS X-Ray for request tracing.
With this, we are able to leverage the default sampling configuration
from AWS X-Ray (1 request/sec + 5%) and integrate it into API Gateway
and our Lambda Functions.

This configures API Gateway to integrate with X-Ray and allows us to
then follow the request(s) into Lambda to see performance information
including initialization and overhead. We also wrap the DynamoDB and SQS
API clients to profile the perfomance of those endpoints as well (this
is natively supported by the X-Ray SDK).