Update Date handling logic
The AWS API Gateway service only supports JSON Schema Draft 4 for object
schemas (not the OpenAPI schemas). This means that `date`, which was not
available as a `format` until later drafts, is not available in API
Gateway. Instead, we need to rely on our ability to parse regular
strings as dates. Because we were using an invalid format, request
validation was failing. It is unfortunate that this error wasn't caught
during the import of the spec into API Gateway.

This also fixes some potential bugs around invalid dates in our
implementation of `isValidDate` and adds some more tests to handle those
edge cases. It very well may introduce more bugs or also leave behind
some subtle ones. Because of that, I have added a TODO to switch to a
trusted and reputable library for this; however, performing that
evaluation was somewhat out-of-scope for this change, which is intended
to be a quick fix to unblock the UI team.