Implement functions to send and pull from the queue
WIP: needs some more work on the functions and tests

Add a function to place messages on the queue and quickly respond
to the client.

Add a function that pulls the messages from the queue to process
emails using nodemailer with secrets from secrets manager.

Add a client for secrets manager in the utils folder.

Ticket: AT-6735