Implement findAdministrators func for business rules
The 'Team Member Roles' business rules require that an
applicationStep has acceptable admin roles at specific levels
The findAdministrators func verifies if the applicationStep
statisfies one or more of the acceptable combinations for admin
roles and returns an object consolidating that information. If the
admin roles are not acceptable, application and environment indices
are provided of where admin roles are missing.

Add unit tests for findAdministrators func.

Ticket: AT-6723