Add a 400 error for NO_SUCH_PORTFOLIO_DRAFT
Add an additional 400 error to errors.ts for when a portfolio
draft is not found or does not exists for GET operations in the API
spec. A similar 404 error is used when the particular step is not
found for an already created Portfolio Draft. The name ends with
the status code to help differentiate the two.

Update imports across the files using the previous name.

Ticket: AT-6723