Use an Aspect to set Removal Policy
Currently, disjoint mechanisms are used to set the removal policy for
constructs across the application. This uses an Aspect to consistently
apply a removal policy whenever we're dealing with a sandbox
environment. For resources in more permanent environments, the default
retention policies are used. In general, the defaults are generally to
retain resources that hold data such as S3 buckets, DynamoDB tables, and
OpenSearch Domains; however, we need to delete those in our Sandbox
environments. This ensures the proper policy is applied to all resources
in those sandboxes while leaving the default for dev/staging/prod, etc.

While this does now special case "dev", "staging", and "prod", which I
am not a fan of, it resolves the issue of "dev" currently having removal
policies set to DESTROY incorrectly. The same TODO of needing to
determine a long-term fix for this still applies; new tech debt is not
added, it's just moved.