Update state machine validation func and comment
Update validateCompletePortfolioDraft using middy and imported
schema to check for validation. If validation succeeds  the
input enters into the lambda function logs the input and returns
and updates the validation on the portfolio draft. If validation
fails an error middleware updates the validation on the portfolio
draft.

Add comment in web-api-stack regarding the functions used with the
state machine.

Ticket: AT-6743