Update state machine for provisionig requests
Define the structure of the state machine and tasks. Define the
inputs and outputs moving from each state to perform the next
action in the state machine.

Update submitQueue consumer func to better model the shape of the
provisioning job entering into the state machine.

Update persistPortfolioDraft func due to change in the input.

Update unit tests.

Ticket: AT-6844