troubleshooting unit tests with ipLogging
The initial failing of the tests was because of an extra line
for linting, which I removed.

After fixing the linting a lot of the unit tests failed for the
various steps. I did a little troubleshooting and it appears to be
due to missing properties on the 'request.event' property for the
middleware.

I added a sample in createPortfolioStep.test.ts to demonstrate a
fix, but there is probably a better way.