Bump default memory size
This improves peformance. In some quick checks, we see results come back
significantly faster with 256MB memory instead of 128MB (by a factor of
at least 3x; we get timeouts at 128MB and execute in <1s with 256mB). We
probably need to do more serious performance analysis on these functions
but this is a good start.