Update convertSchema fn to support both api yaml files...
Because the internal and provisioning api have slightly different models,
we will need a separate schema.json file. This is a quick fix, which will be removed
once we completely phase out the old API.

Named the new schema "internalSchema" so we don't have to update the old imports, we can
change the name to "schema.json" once the old API is phased out.

Due to the packageRoot file living in the infrastructure/utils folder, we can't pass the route
and will have declare a convert schema for provisioning and internal.