Update to AWS CDK v2
v2 of the AWS CDK has been release. The required changes were primarily
around updating our dependencies and imports. No experimental/alpha
constructs are being used, so we don't have to pull in any packages
beyond @aws-cdk/assert, aws-cdk-lib, and constructs. This should make
future dependency management far easier.

We were already using modern values in `cdk.json`'s `context` block so
we don't need to worry about that. We may need to re-run `cdk bootstrap`
in order to use the newer version of that but even that seems to be
optional.