Implement createEnvironment func
Add an environment into the RDS postgres database to be used with
the new 'atat_internal_api' spec. This is a part of the switch
from DynamoDB to RDS.

New models used with the RDS database will be placed in an
'orm' package with the entities for a clean separation of new
models from the old models. Each requestBody has a slightly
different shape depending on the operation. Additional interfaces
are being added to the Environment entity file as the interfaces are
related to the operations on the Environment table.

Add validation for new pathParameter. Use EnvironmentRepository for
creating an environment.

Provide alternative schema without using 'allOf'.

Ticket: AT-6891