Update and move createEnvironment.ts
Update createEnvironment to include new interfaces and move
location to the 'api/portfolios/environments/' folder since the
operation is related to the new internal portfolios api. The
createConnection function contains the 'ormconfig.json' contents
to be able to test with the local database and can be removed
once using a deployed RDS database.

Update EnvironmentRepository to include new interfaces. Move
location of EnvironmentRepository to be inside of the
'packages/api' to prevent errors such as 'RepositoryNotFoundError'.

Ticket: AT-6891