Return an empty group list when unable to parse
If there is not a claim for groups or if we have the wrong key, rather
than failing to generate a token at all, we should just return an empty
list of groups.

To make logs somewhat easier to query in CloudWatch, I've made them
semi-structured. Clearly we still need to decide our logging framework
but I wanted to play with this for now.

This avoids an issue we're experiencing with a recently-added IdP that
either is not providing a groups attribute or the attribute name is
different than what we expect.