Refactor createEnvironment error handling
Remove catch block to allow thrown errors to be captured by the
errorHandlingMiddleware. This keeps the errors in one location.
The error handling for createEnvironment is experimental, and
can be improved. The use of createError provides a way to pass
along different properties with the error or create new ones,
which can be used to unify how the errors are handled in
errorHandlingMiddleware.

Update interfaces in EnvironmentRepository.

Ticket: AT-6891