Add an updated validateRequestShape func
Add a validateRequestShape func in a new file to provide basic
validation for the requests sent using the 'atat-internal-api'
spec. The new internal api spec requires a slight change in the
path parameters and portfolioDraftId is no longer being used.

All of the pathParameters are being validated by API Gateway, so
there is no need to include validation for the required
pathParameters (e.g., portoflioId, environmentId, applicationId).