Update schema to remove 'allOf' for Environment object
Remove all the properties from the objects in 'allOf' to be
under the properties of the Environment object. This allows the
use of additionalProperties set to false and ensures that there are
no issues with middy failing due to properties nested inside of
'allOf'. Using 'allOf' and 'additionalProperties' is a known issue
when composing schema objects.

- [Schema Composition with allOf](https://json-schema.org/understanding-json-schema/reference/combining.html#properties-of-schema-composition)

Ticket: AT-6891