Refactor updateEnvironment and createEnvironment funcs
Refactor and consolidate logic to check for uniqe names. Add
uniqueNameValidator func to allow reuse when checking for unique
names at various endpoints. Update error handling for duplicate
names to be more generic. Refactor createEnvironment to use
uniqueNameValidator.

Refactor logic to update environment to be contained in the
updateEnvironment method in the EnvironmentRepository. The method
updates and makes a call to get the updated environment so that
the lambda only needs to call one method instead of two.

Ticket: AT-6893