Resolve middleware validator issues....
Middy validator is built on ajv, and it has problems with using allOf and
additionalProperties=false. We resolve those issues by wrapping the schema
correctly in the desired shape, we also resolve the issue of being able to
send an id, createdAt, updatedAt, and other database typeorm variables by
wrapping the schema this way.