Add secret scanning with `detect-secrets`
This adds base secret scanning capabilities using the
`Yelp/detect-secrets` project which we've used before. For now, a
`.pre-commit-config.yaml` is added which we will require using as part
of AT-7100 (but it's added now for convenience and to demonstrate that
the selected secret scanning tool works with pre-commit).

Going forward, secrets can be added to the `.secrets.baseline` file;
however, modifying that file requires review from a member of the
`@dod-ccpo/atat-code-maintainers` team.

Additionally secrets going forward can be added to the allowlist in the
baseline file by running detect-secrets scan.

Ticket: AT-7099