Implement provisioning job func
Add a startProvisionJob function with a unit test that takes
a provisioning request from SNOW, translate the request to
a CspInvocation request, and start execution of the
provisioning stateMachine. Code related to responses, errors,
and aws-sdk was resued.

Middy used for validation similar to previously with a
manually created schema. Also, resused code for middleware.
'esModuleInterop' was added to the tsconfig.json file to prevent
an error when running the unit test
("TypeError: (0 , core_1.default) is not a function)").

A models folder containing interfaces and schema files was added,
but this is meant mainly as a starting point, but can be changed
if the team wants to take a different approach to avoid duplication
of the specification, interface, and schema in the codebase.

Ticket: AT-7050