Add state machine infratructure
Add constructs to be used for building state machine workflow.
A sample func as a playholder func inside of the state machine.
Add a state machine to the atat-web-api-stack.

Attempted to reduce code used in the atat-web-api-stack for the
state machine by moving the code into constructor files.

Ticket: AT-7050