Refactor CSP information into a class
Implement feedback to refactor CSP information to ensure CSP
details are correclty linked to each CSP and simplify code.
Unable to use the refactored CloudServiceProvider class for the
'ProvisionRequest' interface which appears to be beasue the
properties are private and can not be used with the interface.

Update files using the CSP information and jest.config.

Ticket: AT-7050