Refactor provisioning state machine components
Attempted to refactor the state machine components to reduce
the amount of code in the atat-web-api-stack.

Most of the code was consolidated into one file as they are all
related and used for the same state machine. The
'ProvisioningWorkflow' class was created as an approach to
simplify the tasks and funcs required for the state machine
definition. Many of the components depend on one another and
require reference to the stack scope, which makes it a bit
difficult to cleanly seperate and group things.

This is not considered to be optimal, but a start and there is
room for improvement.

Ticket: AT-7050