Update IAM configuration to work with CDKv2 roles
The main changes to actually add the necessary permissions are in
`lib/atat-iam-stack.ts`. The remaining changes are simply adding
additional support code to make it easier to manage the deployments.

Eventually (but soon) we probably want to talk about identifying a path
to removing the `atat:EnvironmentId` context value as a requirement for
deploying the base infrastructure (since we really only want there to be
one per _account_ not per environment; and in some cases
account:environment may either be 1:1 or 1:N).

The change to `.secrets.baseline` is because the line moved within the
file.