Update sfn workflow
Add mock invocation function to the workflow. Update the
InvokeCspApi task to retry when a custom 500 erorr is thrown and
catch for all other errors. This simplifies and allows for the task
to handle retries based on the error thrown in the lambda. It also
allows for additional tasks to be chained after the choice state.

If the choice state references a state/task that already has a next
state (e.g., InvokeCspApi task), there is an error and
'.afterwards()' cannot be used to chain additional tasks. I believe
this is a limitation seen when forming sfn with CDK, rather than
the sfn itself.

Ticket: AT-7055