Ensure CDK SSM params can be read in all regions
IAM is a global resource. Because of that, in our IAM policies that are
intended to be usable in multiple regions (which the resources in
`atat-iam-stack` are, even though we currently only use a single
region), we need to _not_ specify specific AWS regions for regional
operations. A principal defined in this stack should be able to perform
actions on resources in _any_ GovCloud region.