Update infrastructure to consume queue messages
Update queue to a FIFO queue which requires deduplication to be
enabled and a message group id.

Update APIGW to include a GET /provisioning-job method and a
consume-provisioning-job func.

Update api-sfn-function to make stateMachine optional, with a plan
to refactor.

Remove unused property in provisioning spec, sample-fn, and comments.

Ticket: AT-7058