Refactor to allow deploying full environments
This configures the application pipeline to deploy a "full" environment.
It also allows for building "sandboxes" disconnected from a VPC and the
rest of the environment. There will still be a "Sandbox" environment to
build some guardrails for IAM (and the pipeline itself for that) but
it's really no big deal (hopefully). Ideally, the only developer-facing
change to deploy a sandbox environment would be the need to now add
`-c atat:sandbox=true` to CLI invocations for diff/deploy.