Create deployment pipeline (#739)
Harden the deployment process to use a pipeline

This adopts a pipeline to deploy changes to non-sandbox environments and
solidifies the idea of what is in fact a "sandbox" environment by simply
leveraging a new context value. The pipeline is implemented using the
CDK's native CDK Pipelines construct library (`aws-cdk-lib.pipelines`).
This uses CodePipeline under the hood to perform the deployment.

Through this, IAM is significantly refactored to remove no-longer-needed
values and permissions (and also adds permissions to view the pipeline)
and moves the "base infrastructure" to be managed by the pipeline. This
sort of starts to treat sandbox environments as a special case rather
than a "real" environment; this more accurately represents how our
environments actually look (where sandbox is the only "weird" account).

We actually remove quite a bit of code here. And a lot of the "new" code
is actually just code that has been moved from other places. For the
most part, the Pipeline stack itself and the `bin/` folder are the only
places that had much net-new code besides the README, which has been
entirely rewritten from the ground up. The changes there should likely
be reviewed carefully. Additionally, the steps developers need to take
to test their environments locally have changed; the details for those
changes are in the new README.

A temporary workaround has been put in place to ensure that we don't
break access to API Gateway from ServiceNow or the internet. We have
more work to do for network transit before we can make that work
perfectly.

The changes here also mean we no longer have custom values in our
`.secrets.baseline` (for now).

Implementation details for how we create and manage the GitHub PAT will
not be covered in this repository and (for now) we will keep those
details internal.