Add a basic IdP Configuration
This adds an initial implementation of an internal identity provider
that we can leverage and have trusted by the various CSPs who may
implement the ATAT API specification.

In this configuration, various client applications can be registered
as IdP clients. The initial CSP implementation is performed with Amazon
Cognito; however, a generic interface is also created to allow other
potential implementations. Additionally, the initial supported
application type is based on Lambda functions but theoretically any
wrapper could be created implementing the necessary interface (including
a State Machine, EC2 instance, or any other resource).

The GovCloudCompatibility Aspect is updated to handle various
differences in Cognito between the standard and GovCloud partitions.

The initial client implementation just fetches a token using the client
credentials flow and then does nothing with it. The token type and
expiration are logged (but not the token itself).