Lint on commit
This automatically runs a `npm run lint:fix` when TS, JS, or JSON files
change. This will increase the amount of time it takes for the
pre-commit checks to run but at least your code will automatically be
linted!

This may not fit well into folks workflows and if it doesn't then let's
not merge it.