Refactor IdP Client into a library
This sorta turns our IdP client into more of a private/internal library
that meets our needs. It doesn't rely on anything too fancy. It lets us
pull configuration for the IdP from the environment and turn it into
something useful. Ideally, this means a client could almost just import
`getToken` and call it with no arguments and get something workable in
the majority of use cases. It's not perfect and won't extend too far
beyong client credentials flow but honestly we shouldn't really need
that.