Don't throw axios error on non-2xx status
By default, `axios` will throw an Error if the status code is outside
the 2xx/3xx range. We don't really want that because we already log the
appropriate parts of requests and we're really more interested in the
error response bodies. And we can always throw again if we want to.