Use consistent log attributes and formatting
This moves the logger definition to a shared location, allowing the
various consumers to use a shared logger. This allows a single logger
instance to be used across imported locations (such as within a Function
and within a library module) to try to have as much cohesion as possible
in the available attributes, context information, and formatting.