Invoke CSP ATAT implementations using tokens from IdP (#856)
* Refactor IdP Client into a library

This sorta turns our IdP client into more of a private/internal library
that meets our needs. It doesn't rely on anything too fancy. It lets us
pull configuration for the IdP from the environment and turn it into
something useful. Ideally, this means a client could almost just import
`getToken` and call it with no arguments and get something workable in
the majority of use cases. It's not perfect and won't extend too far
beyong client credentials flow but honestly we shouldn't really need
that.

* Invoke the API

* Format error/success responses from Mock correctly

* Don't throw axios error on non-2xx status

By default, `axios` will throw an Error if the status code is outside
the 2xx/3xx range. We don't really want that because we already log the
appropriate parts of requests and we're really more interested in the
error response bodies. And we can always throw again if we want to.

* Un-deprecate the JWT decode function

So long as we don't export this function and the name is as gross and
awful as it now is, no one should use it for anything other than the one
safe purpose that it's good for.

* Use consistent log attributes and formatting

This moves the logger definition to a shared location, allowing the
various consumers to use a shared logger. This allows a single logger
instance to be used across imported locations (such as within a Function
and within a library module) to try to have as much cohesion as possible
in the available attributes, context information, and formatting.

* Oops. Messed up a rebase/merge somewhere

* Fixup errors in tests

* Let the invocation function run for a few minutes

Theoretically a CSP should respond pretty quickly (either with a fast
200 or a fast 202) but they may not. So we'll let this run for awhile
with room to grow if needed.

* Document manual deployment steps

* Add tests

* It's not a mock anymore!

* Get rid of some unused imports

* Remove unreachable or dead code

* Add tests for csp-write-portfolio

* Cleanup imports

* Remove "Mock" from error name